requirements.txt
colorama==0.4.4
PyYAML==5.4.1
tradingview-ta
kucoin-python